Revert "MB-738 Datepicker improperly handles typing input"
