+++ src/scenes/Office/Ppm/api.js
@@ -1,10 +1,10 @@
 import { getClient, checkResponse } from 'shared/Swagger/api';
-import { formatDateForSwagger } from 'shared/dates';
+import { formatDateString } from 'shared/utils';
 
 export async function GetPpmIncentive(moveDate, originZip, originDutyStationZip, destZip, weight) {
   const client = await getClient();
   const response = await client.apis.ppm.showPPMIncentive({
-    original_move_date: formatDateForSwagger(moveDate),
+    original_move_date: formatDateString(moveDate),
     origin_zip: originZip,
     origin_duty_station_zip: originDutyStationZip,
     destination_zip: destZip,
