+++ src/shared/JsonSchemaForm/SingleDatePicker.jsx
@@ -1,8 +1,36 @@
 import React from 'react';
 import DayPickerInput from 'react-day-picker/DayPickerInput';
-import { parseDate, formatDate, defaultDateFormat } from 'shared/dates';
+import moment from 'moment';
+import { defaultDateFormat } from 'shared/utils';
 
 import 'react-day-picker/lib/style.css';
+
+// First date format is take to be the default
+const allowedDateFormats = [
+  defaultDateFormat,
+  'YYYY/M/D',
+  'YYYY-M-D',
+  'M-D-YYYY',
+  'D-MMM-YYYY',
+  'MMM-D-YYYY',
+  'DD-MMM-YY',
+];
+
+function parseDate(str, _format, locale = 'en') {
+  // Ignore default format, and attempt to parse date using allowed formats
+  const m = moment(str, allowedDateFormats, locale, true);
+  if (m.isValid()) {
+    return m.toDate();
+  }
+
+  return undefined;
+}
+
+function formatDate(date, format = defaultDateFormat, locale = 'en') {
+  return moment(date, allowedDateFormats, locale, true)
+    .locale(locale)
+    .format(format);
+}
 
 const getDayPickerProps = disabledDays => {
   return {
